# Small async perf improvement