--- a/ProtoPromise_Unity/Assets/Plugins/ProtoPromise/Core/Promises/Internal/PromiseFieldsInternal.cs
+++ b/ProtoPromise_Unity/Assets/Plugins/ProtoPromise/Core/Promises/Internal/PromiseFieldsInternal.cs
@@ -461,6 +461,7 @@ namespace Proto.Promises
 
             partial class AsyncPromiseRef<TResult> : AsyncPromiseBase<TResult>
             {
+                private HandleablePromiseBase _nextForComplete;
 #if PROMISE_PROGRESS
                 private float _minProgress;
                 private float _maxProgress;