Changed async thread static field to instance field with Interlocked.CompareExchange. This decreases execution time at the cost of increased memory.

Master:

| Pending |     Mean |     Error |    StdDev | Allocated | Survived |
|-------- |---------:|----------:|----------:|----------:|---------:|
|    True | 2.280 μs | 0.0298 μs | 0.0233 μs |         - |    688 B |

This PR:

| Pending |     Mean |     Error |    StdDev | Allocated | Survived |
|-------- |---------:|----------:|----------:|----------:|---------:|
|    True | 2.182 μs | 0.0165 μs | 0.0154 μs |         - |    712 B |